requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
beautifulsoup4  # Instead of bs4
arxiv
pymupdf
wikipedia-api  # Instead of wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search==5.3.1b1  # Corrected package name
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube-transcript-api
unstructured
pytube
numexpr
huggingface-hub

